+++ src/instagram/core.nim
@@ -2,11 +2,21 @@ import std/asyncdispatch
 from std/uri import parseUri, `$`
 from std/strutils import `%`, contains
 from std/json import parseJson, `{}`, getStr, JsonParsingError
+from std/httpcore import HttpHeaders
+
+import pkg/jsony
 
 import pkg/unifetch
 from pkg/util/forStr import between
 
 const igDebugReqFile {.strdefine.} = ""
+
+# import instagram/parsing
+import types/xigSharedData
+
+proc renameHook*[T: object](v: var T; fieldName: var string) =
+  if fieldName.len > 2 and fieldName[0..1] == "__":
+    fieldName = fieldName[2..^1]
 
 type
   IgTooManyRequests* = object of IOError
@@ -22,9 +32,9 @@ type
 type
   Instagram* = ref object
     appId: string ## X-IG-App-ID
-    csrfToken: string ## X-CSRFToken
     cookies*: string ## Optional, but without authentication there's a rate limit
-    userId*: string ## Logged user ID, gets from cookie
+    sharedData*: XigSharedData
+    headers: HttpHeaders
 
 const
   instagramUrl = parseUri "https://www.instagram.com"
@@ -33,48 +43,63 @@ const
 using
   ig: Instagram
 
+proc parseSharedData(body: string): XigSharedData =
+  ## Parses XIGSharedData from Instagram HTML
+  var json = getStr parseJson "\"" & body.between("raw\":\"", "\",\"native\"") & "\""
+  if json.len == 0:
+    return
+  result = json.fromJson XigSharedData
+
 proc prepare*(ig) {.async.} =
   ## Adds to the provided mutable Instagram instance the required codes for API
   ## request extracted from a Instagram page
   ##
   ## Need to run when `IgMissingCsrf` is raised
   let
-    uni = newUniClient()
+    uni = newUniClient(headers = ig.headers)
     req = await uni.get instagramUrl
     body = req.body
   close uni
 
   ig.appId = body.between("X-IG-App-ID\":\"", "\"")
-  ig.csrfToken = body.between("\\\"csrf_token\\\":\\\"", "\\\"")
 
-  if "user_id" in ig.cookies:
-    ig.userId = ig.cookies.between("user_id=", ";")
+  ig.sharedData = parseSharedData body
 
 proc newInstagram*(cookies = ""): Future[Instagram] {.async.} =
   ## Creates new Instagram instance
   new result
   result.cookies = cookies
+  result.headers = newHttpHeaders({
+    "X-Requested-With": "XMLHttpRequest",
+    # "Alt-Used": instagramUrl.hostname,
+    "Origin": $instagramUrl,
+    "Referer": $instagramUrl,
+  })
+
   if cookies.len > 0:
-    if cookies[^1] != ';':
-      result.cookies.add ";"
-    result.cookies.add "csrftoken="
+    result.headers.add("Cookie", result.cookies)
 
   await prepare result
+
+  result.headers.add("X-IG-App-ID", result.appId)
+
+  if result.sharedData.csrfToken.len > 0:
+    result.headers.add("X-CSRFToken", result.sharedData.csrfToken)
+    if cookies.len > 0:
+      if cookies[^1] != ';':
+        result.cookies.add ";"
+      result.cookies.add "csrftoken=" & result.sharedData.csrfToken
+
+  if cookies.len > 0:
+    result.headers.add("Cookie", result.cookies)
+
 
 # Internal proc
 proc request*(ig; httpMethod: HttpMethod; endpoint: string; body = ""): Future[string] {.async.} =
   ## Requests to Instagram internal api
   if httpMethod == HttpPost and ig.cookies.len < 100:
     raise newException(IgAuthRequired, "In order to proceed, provide your Instagram cookies to `newInstagram` procedure.")
-  let uni = newUniClient(headers = newHttpHeaders({
-    "X-IG-App-ID": ig.appId,
-    "X-CSRFToken": ig.csrfToken,
-    "X-Requested-With": "XMLHttpRequest",
-    # "Alt-Used": instagramUrl.hostname,
-    "Origin": $instagramUrl,
-    "Referer": $instagramUrl,
-    "Cookie": ig.cookies & ig.csrfToken
-  }))
+  let uni = newUniClient(headers = ig.headers)
   if httpMethod == HttpPost:
     uni.headers["Content-Type"] = "application/x-www-form-urlencoded"
 
@@ -104,7 +129,8 @@ func endpoint*(path: string; args: varargs[string, `$`]): string =
   path % args
 
 when isMainModule:
+  import json
   const cookies = staticRead "../../developmentcookies.txt"
   # let ig = waitFor newInstagram ""
   let ig = waitFor newInstagram cookies
-  echo ig[]
+  echo pretty %*ig.sharedData[]